requirements.txt
Flask==0.12.2
python_clementine_remote==1.0.3
SQLAlchemy==1.1.9
paho_mqtt==1.3.1
